fix: create FilterChip component
