+++ packages/admin/dashboard/src/components/table/data-table/data-table-filter/number-filter.tsx
@@ -1,11 +1,10 @@
-import { EllipseMiniSolid, XMarkMini } from "@medusajs/icons"
-import { Input, Label, Text, clx } from "@medusajs/ui"
+import { EllipseMiniSolid } from "@medusajs/icons"
+import { Input, Label, clx } from "@medusajs/ui"
 import * as Popover from "@radix-ui/react-popover"
 import * as RadioGroup from "@radix-ui/react-radio-group"
 import { debounce } from "lodash"
 import {
   ChangeEvent,
-  MouseEvent,
   useCallback,
   useEffect,
   useState,
@@ -15,6 +14,8 @@ import { useTranslation } from "react-i18next"
 import { useSelectedParams } from "../hooks"
 import { useDataTableFilterContext } from "./context"
 import { IFilter } from "./types"
+import { TFunction } from "i18next"
+import FilterChip from "./filter-chip"
 
 type NumberFilterProps = IFilter
 
@@ -40,7 +41,7 @@ export const NumberFilter = ({
   })
 
   const currentValue = selectedParams.get()
-  const [previousValue, setPreviousValue] = useState<string[] | null | undefined>(
+  const [previousValue, setPreviousValue] = useState<string[] | undefined>(
     currentValue
   )
 
@@ -135,12 +136,16 @@ export const NumberFilter = ({
   const LT_KEY = `${key}-lt`
   const EQ_KEY = key
 
+  const displayValue = parseDisplayValue(currentValue, t)
+  const previousDisplayValue = parseDisplayValue(previousValue, t)
+
   return (
     <Popover.Root modal open={open} onOpenChange={handleOpenChange}>
-      <NumberDisplay
-        previousValue={previousValue}
+      <FilterChip
+        hasOperator
+        hadPreviousValue={!!previousDisplayValue}
         label={label}
-        value={currentValue}
+        value={displayValue}
         onRemove={handleRemove}
         readonly={readonly}
       />
@@ -247,7 +252,7 @@ export const NumberFilter = ({
   )
 }
 
-const parseDisplayValue = (value: string[] | null | undefined, t) => {
+const parseDisplayValue = (value: string[] | null | undefined, t: TFunction) => {
   const parsed = JSON.parse(value?.join(",") || "{}")
   let displayValue = ""
 
@@ -273,87 +278,6 @@ const parseDisplayValue = (value: string[] | null | undefined, t) => {
   }
 
   return displayValue
-}
-
-const NumberDisplay = ({
-  label,
-  value,
-  previousValue,
-  readonly,
-  onRemove,
-}: {
-  label: string
-  value?: string[]
-  previousValue?: string[]
-  readonly?: boolean
-  onRemove: () => void
-}) => {
-  const { t } = useTranslation()
-  const handleRemove = (e: MouseEvent<HTMLButtonElement>) => {
-    e.stopPropagation()
-    onRemove()
-  }
-
-  const displayValue = parseDisplayValue(value, t)
-  const previousDisplayValue = parseDisplayValue(previousValue, t)
-
-
-  return (
-    <div className="bg-ui-bg-field transition-fg shadow-borders-base text-ui-fg-subtle flex cursor-default select-none items-stretch rounded-md">
-      {!previousDisplayValue && (
-        <Popover.Anchor />
-      )}
-      <div
-        className={clx("flex items-center justify-center px-2 py-1", {
-          "border-r": !!(previousDisplayValue || displayValue),
-        })}
-      >
-        <Text size="small" weight="plus" leading="compact">
-          {label}
-        </Text>
-      </div>
-      {!!(previousDisplayValue || displayValue) && (
-        <div className="border-r p-1 px-2">
-          <Text
-            size="small"
-            weight="plus"
-            leading="compact"
-            className="text-ui-fg-muted"
-          >
-            {t("general.is")}
-          </Text>
-        </div>
-      )}
-      {!!(previousDisplayValue || displayValue) && (
-        <Popover.Trigger
-          asChild
-          className={clx(
-            "flex cursor-pointer items-center border-r p-1 px-2",
-            {
-              "hover:bg-ui-bg-field-hover": !readonly,
-              "data-[state=open]:bg-ui-bg-field-hover": !readonly,
-            }
-          )}
-        >
-          <Text size="small" weight="plus" leading="compact">
-            {displayValue}
-          </Text>
-        </Popover.Trigger>
-      )}
-      {!readonly && !!(previousDisplayValue || displayValue) && (
-        <button
-          onClick={handleRemove}
-          className={clx(
-            "text-ui-fg-muted transition-fg flex items-center justify-center p-1",
-            "hover:bg-ui-bg-subtle-hover",
-            "active:bg-ui-bg-subtle-pressed active:text-ui-fg-base"
-          )}
-        >
-          <XMarkMini />
-        </button>
-      )}
-    </div>
-  )
 }
 
 const parseValue = (value: string[] | null | undefined) => {
