+++ packages/admin/dashboard/src/components/table/data-table/data-table-filter/string-filter.tsx
@@ -1,12 +1,11 @@
-import { XMarkMini } from "@medusajs/icons"
-import { Input, Label, Text, clx } from "@medusajs/ui"
+import { Input, Label, clx } from "@medusajs/ui"
 import * as Popover from "@radix-ui/react-popover"
 import { debounce } from "lodash"
 import { ChangeEvent, useCallback, useEffect, useState } from "react"
-import { useTranslation } from "react-i18next"
 import { useSelectedParams } from "../hooks"
 import { useDataTableFilterContext } from "./context"
 import { IFilter } from "./types"
+import FilterChip from "./filter-chip"
 
 type StringFilterProps = IFilter
 
@@ -71,8 +70,9 @@ export const StringFilter = ({
 
   return (
     <Popover.Root modal open={open} onOpenChange={handleOpenChange}>
-      <StringDisplay
-        previousValue={previousValue}
+      <FilterChip
+        hasOperator
+        hadPreviousValue={!!previousValue}
         label={label}
         value={query?.[0]}
         onRemove={handleRemove}
@@ -119,86 +119,5 @@ export const StringFilter = ({
         </Popover.Portal>
       )}
     </Popover.Root>
-  )
-}
-
-const StringDisplay = ({
-  previousValue,
-  label,
-  value,
-  readonly,
-  onRemove,
-}: {
-  previousValue?: string
-  label: string
-  value?: string
-  readonly?: boolean
-  onRemove: () => void
-}) => {
-  const { t } = useTranslation()
-
-  return (
-    <div
-      className="bg-ui-bg-field transition-fg shadow-borders-base text-ui-fg-subtle flex cursor-default select-none items-stretch overflow-hidden rounded-md"
-    >
-      {!previousValue && (
-        <Popover.Anchor />
-      )}
-      <div
-        className={clx(
-          "flex items-center justify-center whitespace-nowrap px-2 py-1",
-          {
-            "border-r": !!(value || previousValue),
-          }
-        )}
-      >
-        <Text size="small" weight="plus" leading="compact">
-          {label}
-        </Text>
-      </div>
-      <div className="flex w-full items-center overflow-hidden">
-        {!!(value || previousValue) && (
-          <div className="border-r p-1 px-2">
-            <Text
-              size="small"
-              weight="plus"
-              leading="compact"
-              className="text-ui-fg-muted"
-            >
-              {t("general.is")}
-            </Text>
-          </div>
-        )}
-        {!!(value || previousValue) && (
-          <Popover.Trigger asChild className={clx("flex-1 cursor-pointer overflow-hidden border-r p-1 px-2",
-            {
-              "hover:bg-ui-bg-field-hover": !readonly,
-              "data-[state=open]:bg-ui-bg-field-hover": !readonly,
-            }
-          )}>
-            <Text
-              size="small"
-              leading="compact"
-              weight="plus"
-              className="truncate text-nowrap"
-            >
-              {value}
-            </Text>
-          </Popover.Trigger>
-        )}
-      </div>
-      {!readonly && !!(value || previousValue) && (
-        <button
-          onClick={onRemove}
-          className={clx(
-            "text-ui-fg-muted transition-fg flex items-center justify-center p-1",
-            "hover:bg-ui-bg-subtle-hover",
-            "active:bg-ui-bg-subtle-pressed active:text-ui-fg-base"
-          )}
-        >
-          <XMarkMini />
-        </button>
-      )}
-    </div>
   )
 }
